laundrylink-cloud/requirements.txt
flask==3.1.*
python-dotenv==1.1.*
requests==2.32.*
supabase==2.*
